+++ pkg/sidecar/sidecar_test.go
@@ -184,45 +184,7 @@ func TestSidecar_RunDaemon(t *testing.T) {
 	cancel()
 }
 
-//Tests that when there is no defaultTimeout in the config, it uses
-//the default defaultTimeout set in a constant in the spiffe_sidecar
-func Test_getTimeout_default(t *testing.T) {
-	config := &Config{}
 
-	expectedTimeout := defaultTimeout
-	actualTimeout, err := GetTimeout(config)
-
-	assert.NoError(t, err)
-	if actualTimeout != expectedTimeout {
-		t.Errorf("Expected defaultTimeout : %v, got %v", expectedTimeout, actualTimeout)
-	}
-}
-
-//Tests that when there is a timeout set in the config, it's used that one
-func Test_getTimeout_custom(t *testing.T) {
-	config := &Config{
-		Timeout: "10s",
-	}
-
-	expectedTimeout := time.Second * 10
-	actualTimeout, err := GetTimeout(config)
-
-	assert.NoError(t, err)
-	if actualTimeout != expectedTimeout {
-		t.Errorf("Expected defaultTimeout : %v, got %v", expectedTimeout, actualTimeout)
-	}
-}
-
-func Test_getTimeout_return_error_when_parsing_fails(t *testing.T) {
-	config := &Config{
-		Timeout: "invalid",
-	}
-
-	actualTimeout, err := GetTimeout(config)
-
-	assert.Empty(t, actualTimeout)
-	assert.NotEmpty(t, err)
-}
 
 func TestGetCmdArgs(t *testing.T) {
 	cases := []struct {
